Set a name for sessions
